Added password removal option and auth providers management

+++ auth.ts
@@ -1,14 +1,17 @@
 import NextAuth from "next-auth";
-import authConfig from "./auth.config";
+import authConfig from "@/auth.config";
 import CredentialsProvider from "next-auth/providers/credentials";
 import { PrismaAdapter } from "@auth/prisma-adapter";
 import { UserRoles } from "@prisma/client";
 import {
 	findUserByEmail,
 	findUserById,
+	getCurrentAuthUser,
 	matchPassword,
-} from "./app/api/actions/user";
+} from "@/app/api/actions/user";
 import db from "@/lib/db";
+import { parseUsername } from "@/lib/user";
+import { cookies } from "next/headers";
 
 declare module "next-auth" {
 	interface User {
@@ -22,16 +25,47 @@ export const { handlers, auth, signIn, signOut } = NextAuth({
 	callbacks: {
 		// biome-ignore lint/suspicious/noExplicitAny: <explanation>
 		async signIn({ user, account }): Promise<any> {
-			user.userName = user.id;
+			user.userName = parseUsername(user.id);
 
-			// Delete any existing provider with the same id
-			// Ideally it shouldn't happen, but it may happen if a user was deleted from the database but his provider accounts were not deleted
-			await db.account.deleteMany({
-				where: {
-					provider: account.provider,
-					providerAccountId: account.providerAccountId,
-				},
-			});
+			const alreadyExistingUser = await getCurrentAuthUser();
+
+			if (alreadyExistingUser?.id) {
+				const existingProviders = await db.account.findMany({
+					where: {
+						userId: alreadyExistingUser?.id,
+						provider: account.provider,
+					},
+				});
+
+				if (existingProviders?.length > 0) {
+					try {
+						await db.account.deleteMany({
+							where: {
+								userId: alreadyExistingUser?.id,
+								provider: account.provider,
+							},
+						});
+					} catch (error) {}
+				} else if (
+					existingProviders?.at(0)?.userId &&
+					alreadyExistingUser?.id !== existingProviders?.at(0)?.userId
+				) {
+					return null;
+				}
+
+				return alreadyExistingUser;
+			}
+
+			try {
+				await db.account.deleteMany({
+					where: {
+						userId: user?.id,
+						provider: account?.provider,
+					},
+				});
+			} catch (error) {
+				console.log({ error });
+			}
 
 			return user;
 		},
